fix(convert): store numeric int8 genotype columns in parquet/feather matrices
The matrix encoders produce string tokens ('0'/'1'/'2'/'3'/'-1'), which
pandas writes to parquet/feather as string columns. The train stage then
loads an all-object feature frame and XGBoost aborts with
'DMatrix.dtypes must be int, float, bool or category', dumping every
column name into the log. CSV output was unaffected (pandas infers
numeric on read), which is why the bug only shows on the default
--out-format parquet path.

Cast the frame to int8 in _write_matrix for binary formats; dosages are
small integers (-1/0/1/2/3) so int8 is lossless. Update the perf tests
to assert the numeric on-disk contract.

Found by the cucumber end-to-end pipeline test (tests/pipeline/).

+++ tests/test_convert_perf.py
@@ -46,7 +46,8 @@ def test_convert_to_matrix_vectorized_encoding_012(tmp_path):
     assert df.loc["s2"].tolist() == ["2", "1", "0", "1"]
 
     on_disk = pd.read_parquet(out_file).set_index("ID")
-    pd.testing.assert_frame_equal(on_disk, df)
+    # Binary formats store genotype dosages as numeric (int8) columns.
+    pd.testing.assert_frame_equal(on_disk, df.astype("int8"))
 
 
 def test_convert_to_matrix_vectorized_encoding_centered(tmp_path):
@@ -91,7 +92,8 @@ def test_vcf_numeric_to_matrix_collects_df(tmp_path):
     assert df.loc["s2"].tolist() == ["1", "3"]
 
     on_disk = pd.read_parquet(out_file).set_index("ID")
-    pd.testing.assert_frame_equal(on_disk, df)
+    # Binary formats store genotype dosages as numeric (int8) columns.
+    pd.testing.assert_frame_equal(on_disk, df.astype("int8"))
 
 
 def _run_ped_map_workflow(tmp_path, threads=1):
